backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
email-validator==2.1.1
